Create the pipeline in the C program
Do the same as cfcr-gst-launch-record.sh, but in C.

From the gst-launch manpage:
"Please  note  that  gst-launch  is  primarily  a debugging tool for
developers and users. You should not build applications on top of it."

And anyway, gst-launch is not flexible enough for our purpose (we need
to listen to external start/stop signals to do the recording at the
right time, for the "offline" solution).