Start/stop video recording when receiving ZeroMQ requests
This requires to copy a private header from libcheese. The API of
libcheese-gtk is not really great, it uses GSettings internally to
choose the camera for example. It is not an API as good as for a real
library.

But it works remarquably well! The resulting code is really
straightforward.