cheese: import all libcheese source code
libcheese is not really a nice library, it's very specific to the GNOME
Cheese application. So I'll modify the libcheese, remove the code that
we don't need, and do small modifications. I think it'll easier to start
with something that works and removing code, than writing something new
from scratch. The libcheese is quite close to what we need.